Add useful shell commands to umdk-lsm6ds3 module

+++ unwired-modules/umdk-lsm6ds3/include/lsm6ds3.h
@@ -109,6 +109,6 @@ bool lsm6ds3_get_raw(lsm6ds3_t *dev, lsm6ds3_data_t *data);
  *
  * @return Temperature in Celsius degrees
  */
-uint16_t lsm6ds3_read_temp_c(lsm6ds3_t *dev);
+int lsm6ds3_read_temp_c(lsm6ds3_t *dev);
 
 #endif /* LSM6DS3_H_ */
